+++ NEWS.md
@@ -11,6 +11,7 @@
 * Support Infinity and NaN.
 * Add strict comparison operators that error out when dimensions of arguments
   do not match: << <<= == <<>> >>= >>.
+* Add range type over units: unitrange.
 
 6.0: Mar 7, 2018
 ----------------